Create & use method for making magic properties consistent
Method added to `Traits\Magic_Methods` so that it is automatically used.

It is protected, but not final. It cannot be called from outside of the
class, but it can be overrided by any class using it to adapt as needed.

All `Traits\Magic\*` call this with the `$prop` given them.