Revert to not using Magic Methods
It is possible (even likely) that a class using this trait will have a
different and non-compatible `MAGIC_POPERTY`. Not as legible, but more
compatible.

Also, better filtering by checking `$components` and existence of URL
property separately where appropriate (do not set `host` to "localhost"
if `host` is not in `$components`).