New trait for highlighting and unhighlighting
Provides two methods: highlightFile & unhighlightString

highlightFile is almost identical to
`highlight_string(file_get_contents())` except that it will convert
colors set through `style` to `<font color...>`. This is for
compatibility with CSP when `style-src` does not inlclude `unsafe-
inline` and may be changed by passing an empty array to `$replacements`

unhighlightString attempts to do the inverse of highlightFile, though
some changes such as tabs are difficult, if not impossible to accurately
revert. By default, this does *not* update the file.